microservices/user_service/requirements.txt
bcrypt==4.0.1
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==9.5.0
PyJWT==2.10.1
python-dotenv==1.0.0
